fixed new Linux warnings

+++ cn/ggraph.cpp
@@ -73,7 +73,6 @@ using paa::KnownGene;
 using paa::KnownGenes;
 using paa::Point;
 using paa::Radio;
-// using Reference = paa::UnMappedReference;
 using paa::Reference;
 using paa::ThreadPool;
 using paa::X11App;
@@ -293,9 +292,9 @@ bool add_genes(const Reference & ref, const CN_abspos & cn_abspos,
   using ChromPos = pair<unsigned int, unsigned int>;
   std::vector<KnownGene>::const_iterator gene_limits[2];
   for (const bool high : {false, true}) {
-    const unsigned int abspos{min(
-        cn_abspos.n_positions(), static_cast<unsigned int>(
-            max(0.0, static_cast<double>(graph.range[0][high]))))};
+    const unsigned int abspos{
+      min(cn_abspos.n_positions(),
+          static_cast<unsigned int>(max(0.0, graph.range[0][high])))};
     const ChromPos chrpos_bound{cn_abspos.chrpos(abspos)};
     gene_limits[high] = upper_bound(
         genes.begin(), genes.end(), chrpos_bound,
